+++ templates/sidebar-rtl/.env.example
@@ -1,0 +1,3 @@
+NUXT_PUBLIC_BASE_API=http://localhost:8000
+NUXT_PUBLIC_BASE_STORAGE=http://localhost:8000
+NUXT_JWT_SECRET=secret_key_here
