+++ templates/sidebar-rtl/app/types/request.ts
@@ -1,11 +1,3 @@
-export interface ApiResponseType<T> {
-  success: boolean
-  status: number
-  message: string
-  data: T
-  redirect_url?: string
-}
-
 export interface ApiFilterResultType<T> {
   entity_count: number
   current_page: number
