--- templates/sidebar-rtl/pnpm-workspace.yaml
@@ -1,8 +1,0 @@
-ignoredBuiltDependencies:
-  - unrs-resolver
-  - vue-demi
-onlyBuiltDependencies:
-  - '@parcel/watcher'
-  - '@tailwindcss/oxide'
-  - esbuild
-  - sharp
