+++ templates/sidebar-rtl/shared/types/request.ts
@@ -1,0 +1,7 @@
+export interface ApiResponseType<T> {
+  success: boolean
+  status: number
+  message: string
+  data: T
+  redirect_url?: string
+}
